chore: cleanup redudant type definitions
